requirements.txt
fastapi==0.109.0
mongoengine==0.27.0
mongoengine_goodjson==1.1.8
passlib==1.7.4
protobuf==4.25.2
pydantic==2.6.0
pymongo==4.6.1
python-dotenv==1.0.1
uvicorn==0.27.0.post1
Werkzeug==3.0.1
